requirements.txt
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
astor=0.7.1=py36_0
binutils_impl_linux-64=2.31.1=h6176602_1
binutils_linux-64=2.31.1=h6176602_7
blas=1.0=mkl
bzip2=1.0.8=h516909a_0
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.5.15=0
cairo=1.16.0=h18b612c_1001
certifi=2019.6.16=py36_0
cudatoolkit=10.1.168=0
cudnn=7.6.0=cuda10.1_0
cupti=10.1.168=0
dbus=1.13.6=he372182_0
expat=2.2.5=he1b5a44_1003
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.0=he983fc9_0
gast=0.2.2=py36_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_7
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
google-pasta=0.1.7=py_0
graphite2=1.3.13=hf484d3e_1000
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_7
h5py=2.9.0=py36h7918eee_0
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=hf484d3e_1000
intel-openmp=2019.4=243
jasper=1.900.1=h07fcdf6_1006
jpeg=9c=h14c3975_1001
keras=2.2.4=py36_1
keras-applications=1.0.7=py_1
keras-preprocessing=1.0.9=py_1
lame=3.100=h14c3975_1001
libblas=3.8.0=10_mkl
libcblas=3.8.0=10_mkl
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libgpuarray=0.7.6=h14c3975_1003
libiconv=1.15=h516909a_1005
liblapack=3.8.0=10_mkl
liblapacke=3.8.0=10_mkl
libpng=1.6.37=hed695b0_0
libprotobuf=3.9.0=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h57b8799_1003
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=h13577e0_1
lz4-c=1.8.3=he1b5a44_1001
mako=1.0.10=py_0
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36h14c3975_0
mkl=2019.4=243
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
opencv=4.1.0=py36h5517eff_4
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1=h7b6447c_0
pcre=8.41=hf484d3e_1003
pillow=6.1.0=py36h34e0f95_0
pip=19.1.1=py36_0
pixman=0.38.0=h516909a_1003
protobuf=3.9.0=py36he1b5a44_0
pthread-stubs=0.4=h14c3975_1001
pygpu=0.7.6=py36h3010b51_1000
python=3.6.8=h0371630_0
pyyaml=5.1.1=py36h7b6447c_0
qt=5.9.7=h52cfd70_2
readline=7.0=h7b6447c_5
scipy=1.2.1=py36h7c811a0_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
sqlite=3.29.0=h7b6447c_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36h3fb9ad6_0
tensorflow-base=1.14.0=gpu_py36he45bfe2_0
tensorflow-estimator=1.14.0=py36h5ca1d4c_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
theano=1.0.4=py36hf484d3e_1000
tk=8.6.8=hbc83047_0
tqdm=4.32.1=py_0
werkzeug=0.15.4=py_0
wheel=0.33.4=py36_0
wrapt=1.11.2=py36h7b6447c_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.8=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.4.0=h3b9ef0a_0
